# api/core.openapi.yaml
basePath: /
definitions:
...
  Project:
    properties:
      created_time:
        type: string
      id:
        type: string
      originated_id:
        description: 'The ID of the project in origination. For example: ACCESS Record
          ID.'
        type: string
      origination:
        description: ACCESS, NAIRR, XRASS, etc.
        type: string
      project_pi_id:
        type: string
      status:
        $ref: '#/definitions/ProjectStatus'
      title:
        type: string
    required:
    - created_time
    - id
    - originated_id
    - origination
    - project_pi_id
    - status
    - title
    type: object
  ProjectListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/ProjectResponse'
        type: array
      total:
        type: integer
    required:
    - items
    - total
    type: object
  ProjectMemberAllocationRef:
    properties:
      id:
        type: string
      name:
        type: string
      role:
        type: string
    required:
    - id
    - name
    - role
    type: object
  ProjectMemberResponse:
    properties:
      added_time:
        type: string
      allocations:
        items:
          $ref: '#/definitions/ProjectMemberAllocationRef'
        type: array
      display_name:
        type: string
      email:
        type: string
      id:
        description: The user's id.
        type: string
      project_id:
        type: string
      role:
        type: string
      status:
        type: string
      user_id:
        type: string
    required:
    - added_time
    - allocations
    - display_name
    - email
    - id
    - project_id
    - role
    - status
    - user_id
    type: object
  ProjectResponse:
    properties:
      created_time:
        type: string
      id:
        type: string
      originated_id:
        description: 'The ID of the project in origination. For example: ACCESS Record
          ID.'
        type: string
      origination:
        description: ACCESS, NAIRR, XRASS, etc.
        type: string
      project_pi_display_name:
        type: string
      project_pi_email:
        type: string
      project_pi_id:
        type: string
      status:
        $ref: '#/definitions/ProjectStatus'
      title:
        type: string
    required:
    - created_time
    - id
    - originated_id
    - origination
    - project_pi_id
    - status
    - title
    type: object
  ProjectStatus:
    enum:
    - ACTIVE
    - INACTIVE
    - DELETED
    type: string
    x-enum-varnames:
    - ProjectActive
    - ProjectInactive
    - ProjectDeleted
...
  updateProjectRequest:
    properties:
      previous_pi_role:
        description: 'The old PI''s role when project_pi_id changes: CO_PI, ALLOCATION_MANAGER
          or MEMBER.'
        type: string
      project_pi_id:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 'REST API for Apache Custos: organizations, users, projects, compute
    clusters, allocations, audit traces, and the privilege/role authorization layer.
    OIDC bearer auth required. Connector endpoints live under /connectors/<name>/.'
  title: Apache Custos Core API
  version: 0.2.0
paths:
...
  /projects:
    get:
      description: Callers without the projects read privilege get only the projects
        they participate in; filters and pagination apply to privileged callers only.
      parameters:
      - description: Filter by PI user ID
        in: query
        name: pi_id
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      - description: Search title / originated_id
        in: query
        name: q
        type: string
      - description: Page size
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProjectListResponse'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List projects (filtered + paginated, PI joined)
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Project payload
        in: body
        name: request
        required: true
        schema:
          properties:
            project_pi_id:
              type: string
            title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Project'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - Projects
  /projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Project has allocations
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a project without allocations
      tags:
      - Projects
    get:
      description: Callers without the projects read privilege only see projects they
        participate in; others return 404.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ProjectResponse'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to change
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/updateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Project'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - Projects
  /projects/{id}/members:
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ProjectMemberResponse'
            type: array
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: List project members (one row per distinct user, with allocations)
      tags:
      - Projects
  /projects/{id}/members/{userId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Role
        in: body
        name: request
        required: true
        schema:
          properties:
            role:
              type: string
          type: object
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: PI change
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set a project member's role; MEMBER removes the role
      tags:
      - Projects
  /projects/{id}/status:
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Status patch
        in: body
        name: request
        required: true
        schema:
          properties:
            status:
              $ref: '#/definitions/ProjectStatus'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Project'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a project's status
      tags:
      - Projects
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"